dom/nodes/moveBefore/tentative/css-transition-to-disconnected-document.html

<!DOCTYPE html>
<title>Node.moveBefore should act like insertBefore when moving to a disconnected document</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
  <section id="old-parent">
    <div id="item"></div>
  </section>
  <section id="new-parent">
  </section>
  <style>
    #item {
      width: 100px;
      height: 100px;
      background: green;
      transition: left 10s;
      position: absolute;
      left: 10;
    }

    section {
      position: relative;
    }

    body {
      margin-left: 0;
    }
  </style>
  <script>
    promise_test(async t => {
      const item = document.querySelector("#item");

      assert_equals(item.getBoundingClientRect().x, 0);
      item.style.left = "400px";
      while (item.getBoundingClientRect().x === 0) {
        await new Promise(resolve => requestAnimationFrame(() => resolve()));
      }
      const doc = document.implementation.createHTMLDocument();
      doc.body.moveBefore(item, null);
      await new Promise(resolve => requestAnimationFrame(() => resolve()));
      assert_equals(item.getBoundingClientRect().x, 0);
    });
  </script>
</body>
